Bump sparkle + move story
